# docs/openapi/api-v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lurus API V2 - Multi-Tenant Gateway",
...
    "version": "2.0.0",
    "contact": {
      "name": "Lurus Support",
      "email": "support@lurus.cn"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://lurus.cn/terms"
    }
  },
  "servers": [
    {
      "url": "https://hub.lurus.cn",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v2/{tenant_slug}/projects/{id}/members": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List Project Members",
        "description": "Lists the members (department leads) of a project. Tenant admin only; a project of another tenant is 404.",
        "operationId": "listProjectMembersV2",
        "security": [
          {
            "OIDCBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ProjectMember"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Tenant admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Add Project Member",
        "description": "Adds a user to a project; a user with no tenant role becomes dept_lead. Idempotent. Tenant admin only. Not available in the shared default tenant (403 TENANT_ROLE_FORBIDDEN_IN_DEFAULT).",
        "operationId": "addProjectMemberV2",
        "security": [
          {
            "OIDCBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id"
                ],
                "properties": {
                  "user_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Tenant admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Project or user not found in this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Remove Project Member",
        "description": "Removes a user from a project; a dept_lead left with no projects is demoted. user_id is read from the query string, or from a JSON body.",
        "operationId": "removeProjectMemberV2",
        "security": [
          {
            "OIDCBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantSlug"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Tenant admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Project or user not found in this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OIDCBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token issued by the deployment's configured OIDC provider. Which provider that is, is deploy-time configuration; no specific vendor is part of this contract."
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie for platform admin"
      }
    },
    "parameters": {
      "TenantSlug": {
        "name": "tenant_slug",
        "in": "path",
        "required": true,
        "description": "Tenant identifier (slug). On routes behind TenantSlugGuard the reserved value `~` means \"the tenant this session authenticated into\"; the server resolves it from the session, and it is what the console sends. An explicit slug must equal that tenant (403 TENANT_MISMATCH otherwise). `GET /user/me` reports the real slug as `tenant_slug`.",
        "schema": {
          "type": "string",
          "pattern": "^([a-z0-9-]+|~)$",
          "example": "lurus"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Invalid request"
          }
        }
      },
...
      "ProjectMember": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "tenant_role": {
            "type": "string"
          },
          "added_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}